==> 16S/R_analyses/14_Coefficients.Rmd <==
---
title: "model coeff."
author: "SRomano"
output: html_document
editor_options: 
  chunk_output_type: console
---

# Load packages

```{r setup}
library(phyloseq)
library(vegan)
library(dplyr)
library(tidyverse)
library(SIAMCAT)

# load my function to run siamcat
source("/Scripts/Model_coefficient_extraction.r")
set.seed(56987)

```

# Load metadata
```{r}
meta<-readRDS("/16S/R_analyses/RDS/meta.rds")
meta$Study_2<-as.character(meta$Study_2)
```

# Load LOG models
```{r}
load("/16S/R_analyses/RData/16S_Sinlge_5perc_log.RData")

lridge<-coefficient.extraction(studies = unique(meta$Study_2) %>% as.character(),
                                list.models = ridge_5x10,
                                model.parallel = T,
                                string.to.add = "log_ridge_5x10",
                                select = "mean.weight")
lridge<-lapply(lridge, function(x) mutate(x, relative.weigths= x$overall.average / sum(abs(x$overall.average))))
lapply(lridge, dim) # different taxa number in different studies, results of taxa filtration.

df1<-do.call(rbind, lridge) %>% as.data.frame()

```

# Format data for figure
```{r}
# add a dummy variable used for formatting
df1$dummy<-rep(1, nrow(df1))

# add another dummy variable to indicate if it is pos or neg
df1$pos.neg<-ifelse(df1$relative.weigths > 0, "POS", "NEG") %>% as.factor()


# make a df with average and std of relative.weights
taxa.weights<-df1 %>%
  group_by(taxa) %>%
  summarise(average = mean(relative.weigths),
            stdev = sd(relative.weigths)) %>%
  as.data.frame()
taxa.weights<-taxa.weights[order(taxa.weights$average, decreasing = T),]
taxa.weights$dir<-ifelse(taxa.weights$average > 0 , "HC", "PD")

saveRDS(taxa.weights, "/16S/R_analyses/RDS/Ridge_taxa.weights.rds")

# add average
df.pos.neg<-df1 %>% 
   group_by(taxa, pos.neg) %>%
   summarise(dummy2 = n()) %>%
   as.data.frame()

for(i in 1:nrow(df.pos.neg)){
  df.pos.neg$average[i]<-taxa.weights[taxa.weights$taxa == df.pos.neg$taxa[i], "average"]
  df.pos.neg$stdev[i]<-taxa.weights[taxa.weights$taxa == df.pos.neg$taxa[i], "stdev"]

}
saveRDS(df.pos.neg, "/16S/R_analyses/RDS/Ridge_df.pos.neg.rds")

```

# Perform ordination
```{r}

df<-reshape2::dcast(df1,
                     study ~ taxa, value.var = "relative.weigths")
# format df
df[is.na(df)]<-0
rownames(df)<-df$study
df<-df[,-1]

dist<-vegdist(df, method = "canberra")
pc<-metaMDS(dist)
# check stress
stressplot(pc)
# looks acceptable
d<-pc$points[] %>%
  as.data.frame()
d$study<-gsub("_.*", "", rownames(d))
d$var<-rownames(d)

#format the meta

meta.short<-meta[!duplicated(meta$Study_2),]
# change Petrov with Europe
meta.short[1,3]<-"Europe"
d<-d[order(d$study),]
meta.short<-meta.short[order(meta.short$Study_2),]
d[which(d$study == "Heintz"),3]<-"Heintz_Bushart"
all(meta.short$Study_2 == d$study)

# merge
d.m<-cbind.data.frame(d, meta.short)
adonis2(dist~Continent, data =  meta.short)

nb.cols <- 24
mycolors <- colorRampPalette(brewer.pal(8, "Set2"))(nb.cols)

# adjust study names
d.m$Study_2<-as.character(d.m$Study_2)
d.m$Study_2[d.m$Study_2 == "Aho"]<-"Aho_2019"
d.m$Study_2[d.m$Study_2 == "Petrov"]<-"Petrov_2017"
d.m$Study_2[d.m$Study_2 == "Keshavarzian"]<-"Keshavarzian_2015"
d.m$Study_2[d.m$Study_2 == "Zhang"]<-"Zhang_2020"
d.m$Study_2[d.m$Study_2 == "Nishiwaki"]<-"Nishiwaki_2020"
d.m$Study_2[d.m$Study_2 == "Hopfner"]<-"Hopfner_2017"
d.m$Study_2[d.m$Study_2 == "Weis"]<-"Weis_2019"
d.m$Study_2[d.m$Study_2 == "Cirstea"]<-"Cirstea_2020"
d.m$Study_2[d.m$Study_2 == "Tan"]<-"Tan_2021"
d.m$Study_2[d.m$Study_2 == "Kenna"]<-"Kenna_2021"
d.m$Study_2[d.m$Study_2 == "Wallen251"]<-"Wallen251_2021"
d.m$Study_2[d.m$Study_2 == "Wallen151"]<-"Wallen151_2021"
d.m$Study_2[d.m$Study_2 == "Jo"]<-"Jo_2022"
d.m$Study_2[d.m$Study_2 == "Lubomski"]<-"Lubomski_2022"
d.m$Study_2[d.m$Study_2 == "Heintz_Bushart"]<-"Heintz_Buschart_2017"
d.m$Study_2[d.m$Study_2 == "Qian"]<-"Qian_2018"
d.m$Study_2[d.m$Study_2 == "Pietrucci"]<-"Pietrucci_2019"


g16<-ggplot(d.m, aes(MDS1, MDS2)) +
  geom_point(aes(color = Study_2), size = 5, stroke = 3, pch = 21) +
  scale_color_manual(values = mycolors) +
  geom_point(aes(fill = Continent), size = 5, stroke = 0, pch = 21) +
  scale_fill_manual(values = c("#8C510A", "#F6E8C3", "#C7EAE5", "#01665E" )) +
  theme_bw() +
  theme(legend.text = element_text(size = 16),
        legend.title = element_text(size =18),
        axis.title = element_text(size =16),
        axis.text = element_blank(),
        plot.caption = element_text(size =16)) +
  labs(color = "Study") 

# load metaG
d.mg<-readRDS("/metaG/R_analyses/Taxonomy/RDS/coeff_nmds.rds")

nb.cols <- 18
mycolors <- colorRampPalette(brewer.pal(8, "Set2"))(nb.cols)
d.mg$study<-gsub("_log.*", "", d.mg$var)

g.smg<-ggplot(d.mg, aes(MDS1, MDS2)) +
  geom_point(aes(color = study), size = 5, stroke = 3, pch = 21) +
  scale_color_manual(values = mycolors) +
  geom_point(aes(fill = Continent), size = 5, stroke = 0, pch = 21) +
  scale_fill_manual(values = c("#8C510A", "#F6E8C3", "#C7EAE5", "#01665E" )) +
  theme_bw() +
  theme(legend.text = element_text(size = 16),
        legend.title = element_text(size =18),
        axis.title = element_text(size =16),
        axis.text = element_blank(),
        plot.caption = element_text(size =16)) +
  labs(color = "Study") +
  guides(fill = "none")


coeff<-cowplot::plot_grid(g16, g.smg,
                   ncol = 2,
                   align = "h",
                   axis = "bl",
                   rel_widths =c(1.1,1),
                   rel_heights = c(1,1),
                   labels = c("auto"), 
                   label_size = 20)

cowplot::save_plot("/16S/R_analyses/Figures/Coeff_cont.svg", coeff, 
                   base_width = 18, base_height = 6)
```
