==> 16S/R_analyses/15_DA_16S.Rmd <==
---
title: "DA"
author: "SRomano"
output: 
  html_document: 
    toc: yes
editor_options: 
  chunk_output_type: console
---

# Load packages

```{r setup}
list.of.packages <- c("dplyr", "ggplot2", "RColorBrewer", "meta", "genodds")
new.packages <- list.of.packages[!(list.of.packages %in% installed.packages()[,"Package"])]
if(length(new.packages)) install.packages(new.packages)
if (!requireNamespace("BiocManager", quietly = TRUE)){
    install.packages("BiocManager")}


library(dplyr)
library(genodds)
library(meta)


set.seed(9821)
```

# Load the data 

This analysis needs to run on a named list in which each item is the abundance table of a study.
```{r}

l_df<-readRDS("/16S/R_analyses/RDS/l.df_g_TSS.rds")
meta<-readRDS("/16S/R_analyses/RDS/meta.rds")

meta %>%
  group_by(Study, PD) %>%
  summarise(n = n()) %>%
  print( n = 50)

dim(meta)
lapply(l_df, dim)

#check the order is the same in all
x<-unlist(lapply(l_df, function(x) names(x)))
all(sort(x) == meta$SampleID)
```

# Format data and meta
```{r}
# rename to use with the genodd function
names(meta)[1]<-"Sample_ID"
names(meta)[15]<-"Cohort"

meta$PD<-factor(meta$PD, levels = c("HC", "PD"))
# I need to transpose it for the DA
l_df<-lapply(l_df, function(x) as.data.frame(t(x)))
# Double check that the TSS is calculate correctly
stopifnot(lapply(l_df, rowSums) %>% unlist() %>% sum() == nrow(meta))
lapply(l_df, rowSums) %>% unlist() %>% sum() == nrow(meta)
```
# Filter
```{r}
spec<-readRDS("/16S/R_analyses/RDS/spec_5x10.rds")

l_df<-lapply(l_df, function(x) x[,which(names(x) %in% spec)])
```

# Calculate DA using GenOdds
```{r}
# source the function
source("/Scripts/DA_function_gen_odd.R")

meta<-meta[order(meta$Sample_ID),]
meta$Cohort<-as.character(meta$Cohort)
l_df<-lapply(l_df, function(x) x[order(rownames(x)),])

names(l_df)[10]<-"Heintz_Bushart" # to make names consistent
rownames(meta)<-meta$Sample_ID

df.cross<-calcualte_da_genOdds(list.df = l_df,
                meta = meta,
                treat = "PD", 
                assume_no_effect = TRUE)
saveRDS(df.cross, "/16S/R_analyses/RDS/df.cross.rds")
```

# Perform meta
```{r}
bins<-unique(df.cross$bin)

df<-data.frame(TE.random = vector(),
                 seTE.random = vector(),
                 lower.random = vector(),
                 upper.random = vector(),
                 pval.random = vector(),
                 bins = vector())
  
for(b in 1:length(bins)){
  require(meta)
    tmp.bin<-df.cross[df.cross$bin == bins[b],]
    print(b)
        
    print(dim(tmp.bin))

    x<-try(met<-metagen(TE=estimate, seTE=std.error,data=tmp.bin, method.tau = "REML",title =bins[b], studlab = Cohort))
    if(class(x)[1] != "try-error"){
    met.df<-as.data.frame(c(met[c("TE.random","seTE.random", "lower.random", "upper.random", "pval.random")],bins= bins[b])) 
    df<-rbind.data.frame(df, met.df)  
    } else { tmp<-data.frame(TE.random =NA, seTE.random = NA, lower.random = NA, upper.random = NA, pval.random = NA, bins= bins[b])
      df<-rbind.data.frame(df, tmp)}
}

df$p.adjust<-p.adjust(df$pval.random, method = "fdr")
```

# Subset sign and save
```{r}
subset(df,p.adjust<0.05 ) %>%
  dim()

# save df with sign only
df.sign<-subset(df,p.adjust<0.05 )
df.sign$dir<-ifelse(df.sign$TE.random<0, "CTRL", "PD")
saveRDS(df.sign, "/16S/R_analyses/RDS/df.sign.rds")

```
