==> 16S/R_analyses/17_Sex_Age_cov.rmd <==
---
title: "Sex_Age_16S"
output: html_document
editor_options: 
  chunk_output_type: console
---
# Load packages and meta

Sex and Age metavariables for the study of Qian et al 2018 need to be obtained from the authors of the original study.

```{r}
library(tidyverse)
library(nlme)
# read medata
meta<-readRDS("/16S/R_analyses/RDS/meta.rds")
```


# Define function
```{r}
lm.conf<-function(microbiome.feats = log_kop,
                  metadata = meta.all,
                  select.desc = "descr"){
    
  df.estimates<-data.frame(feats = vector(length = ncol(microbiome.feats)),
                   
                   p.value.sex.age.PD = vector(length = ncol(microbiome.feats)),
                   est.sex.age.PD = vector(length = ncol(microbiome.feats)),
                  
                   p.value.PD = vector(length = ncol(microbiome.feats)),
                   est.PD =  vector(length = ncol(microbiome.feats)),
                   
                   p.value.sex = vector(length = ncol(microbiome.feats)),
                   est.sex =  vector(length = ncol(microbiome.feats)),

                   p.value.age = vector(length = ncol(microbiome.feats)),
                   est.age =  vector(length = ncol(microbiome.feats)),

                   description =  vector(length = ncol(microbiome.feats)))
    l.mod.sapd<-vector(mode = "list", length = ncol(microbiome.feats))
    l.mod.pd<-vector(mode = "list", length = ncol(microbiome.feats))
    stopifnot(all(rownames(microbiome.feats) == metadata$SampleID))

    for(i in 1:ncol(microbiome.feats)){
      id<-names(microbiome.feats)[i]

      print(paste0(i, " of ", ncol(microbiome.feats)))
      tmp<-microbiome.feats[,id]
      stopifnot(all(rownames(tmp) == metadata$SampleID))
      tmp<-cbind.data.frame(bins = tmp, metadata)
      tmp$age<-as.numeric(tmp$Age)
      tmp$sex<-as.factor(tmp$gender)
      tmp$PD<-as.factor(tmp$PD)
      tmp$Study_2<-as.factor(tmp$Study_2)

      mod<-nlme::lme(bins ~ gender + Age + PD, random = ~ 1 | Study_2,
               data = tmp, na.action = "na.exclude")
      
      l.mod.sapd[[i]]<-mod
      names(l.mod.sapd)[i]<-id

      df.estimates$feats[i]<-id
      df.estimates$p.value.sex.age.PD[i]<-as.data.frame(summary(mod)$tTable)$`p-value`[4]
      df.estimates$est.sex.age.PD[i]<-as.data.frame(summary(mod)$tTable)$Value[4]

      df.estimates$p.value.sex[i]<-as.data.frame(summary(mod)$tTable)$`p-value`[2]
      df.estimates$est.sex[i]<-as.data.frame(summary(mod)$tTable)$Value[2]

      df.estimates$p.value.age[i]<-as.data.frame(summary(mod)$tTable)$`p-value`[3]
      df.estimates$est.age[i]<-as.data.frame(summary(mod)$tTable)$Value[3]

      mod<-nlme::lme(bins ~ PD, random = ~ 1 | Study_2,
               data = tmp, na.action = "na.exclude")
      
      l.mod.pd[[i]]<-mod
      names(l.mod.pd)[i]<-id

      df.estimates$feats[i]<-id
      df.estimates$p.value.PD[i]<-as.data.frame(summary(mod)$tTable)$`p-value`[2]
      df.estimates$est.PD[i]<-as.data.frame(summary(mod)$tTable)$Value[2]

      if(select.desc != "none"){
      df.estimates$description[i] <- df.sign[df.sign$bins == id, select.desc]
      } else {df.estimates$description[i] <- NA}
      
    }
    # adjust pvalues
    df.estimates$p.value.sex.age.PD.adj<-p.adjust(df.estimates$p.value.sex.age.PD, method = "fdr")
    df.estimates$p.value.PD.adj<-p.adjust(df.estimates$p.value.PD, method = "fdr")
    df.estimates$p.value.sex.adj<-p.adjust(df.estimates$p.value.sex, method = "fdr")
    df.estimates$p.value.age.adj<-p.adjust(df.estimates$p.value.age, method = "fdr")

  setClass("lm.df",representation(df="data.frame",l.mod.sapd="list", l.mod.pd="list"))
  data4<-new("lm.df",df=df.estimates,l.mod.sapd=l.mod.sapd,l.mod.pd=l.mod.pd)

  return(data4)
  
}

```

# Run models
```{r}
kop<-readRDS("/16S/R_analyses/RDS/all_g_TSS.rds")
# select samples
dim(kop)
all(meta.all$SampleID %in% names(kop))
kop<-kop[,which(names(kop) %in% meta.all$SampleID)]
dim(kop)

kop<-kop[, meta.all$SampleID]

min<-sort(as.vector(as.matrix(kop)))
min<-min[min!=0]
min<-min[1]/10

# subset sign
df.sign<-readRDS("/16S/R_analyses/RDS/df.sign.rds")
df.sign<-df.sign[order(df.sign$bins),]
kop<-kop[order(rownames(kop)),]

kop<-kop[df.sign$bins,]
all(rownames(kop) == df.sign$bins)

#log trafo
log_kop<-log(kop+min)
log_kop_w<-as.data.frame(t(log_kop))

# order samples
meta.all<-meta.all[order(meta.all$SampleID),]
log_kop_w<-log_kop_w[order(rownames(log_kop_w)),]
stopifnot(all(rownames(log_kop_w) == meta.all$SampleID))

df.estimates<-lm.conf(microbiome.feats = log_kop_w,
                  metadata = meta.all,
                  select.desc = "none") 
saveRDS(df.estimates, "/16S/R_analyses/RDS/Cov.Sex.Age.S4.rds")
df.estimates<-df.estimates@df
```

# Summarise data
```{r}
df.perc<-data.frame(Microbiome.features = c("16S_genera"),
                    prop.conf = vector(length = 1))

b<-dim(subset(df.estimates, p.value.PD.adj < 0.05))
df.estimates$type <- rep("16S_genera", nrow(df.estimates))
df.estimates<-subset(subset(df.estimates, p.value.PD.adj < 0.05), p.value.sex.age.PD.adj >= 0.05)
a<-dim(df.estimates)
df.perc$prop.conf[1]<-round(a[1]/b[1]*100, digits = 1)



# check, format, and save
df.perc$pot.conf<-rep("Sex.Age", nrow(df.perc))
df.perc$seqs<-"16S_amplicons"

df.smg<-rbind.data.frame(
  df.metag<-readRDS("/metaG/R_analyses/Functions/AC_KOpath/RDS/df.perc.pd.drugs_other.meds.rds"),
  df.metag.sa<-readRDS("/metaG/R_analyses/Functions/AC_KOpath/RDS/df.perc.pd.SexAge_bedarf.rds"))
df.smg$seqs<-rep("SMG", nrow(df.smg))

df.all<-rbind.data.frame(df.perc, df.smg)
df.all
write.table(df.all, file =  "/metaG/R_analyses/Functions/AC_KOpath/RDS/df.perc.pd.all.16SSmg.txt", 
            quote = F, 
            row.names = F, 
            col.names = T, 
            sep = "\t")



# select only the sign in lm.PD and not sign in lm.cov.pd
df.estimates<-subset(subset(df.estimates, p.value.PD < 0.05), p.value.sex.age.PD.adj  >= 0.05)


df.conf.form<-data.frame(Microbiome.feature = df.estimates$feats,
                         p.value.sex.age.PD = df.estimates$p.value.sex.age.PD,
                         p.value.sex.age.PD.adj = df.estimates$p.value.sex.age.PD.adj,
                         est.sex.age.PD = df.estimates$est.sex.age.PD,
                         p.value.PD = df.estimates$p.value.PD,
                         p.value.PD.adj = df.estimates$p.value.PD.adj,
                         
                         p.value.sex = df.estimates$p.value.sex,
                         p.value.sex.adj = df.estimates$p.value.sex.adj,
                         est.sex = df.estimates$est.sex,
                       
                         p.value.age = df.estimates$p.value.age,
                         p.value.age.adj = df.estimates$p.value.age.adj,
                         est.age = df.estimates$est.age,
                   
                         est.PD = df.estimates$est.PD,
                         description = rep(NA, nrow(df.estimates)),
                         type = df.estimates$type)

# Combine with metaG
df.metag<-read.delim("/metaG/R_analyses/Functions/AC_KOpath/RDS/Sex.Age.conf.tab_bedarf.txt", header = T)

df.all<-rbind.data.frame(df.metag, df.conf.form)

write.table(df.all, file = "/metaG/R_analyses/Functions/AC_KOpath/RDS/Sex.Age.conf.tab_16S_SMG.txt", 
            quote = F, 
            row.names = F, 
            col.names = T, 
            sep = "\t")

```
